Reindeer positioning is functioning properly.

+++ src/reindeerraces/reindeer/Lane.java
@@ -2,7 +2,7 @@ package reindeerraces.reindeer;
 
 public class Lane
 {
-	public int getValue()
+	public double getValue()
 	{
 		return 0;
 	}
